refactor(core-data): convert ~15 feature services from SwiftData to Core Data (Phase 3B, Batch 1)
Converts Work domain services to use NSManagedObjectContext, NSFetchRequest,
and NSPredicate, replacing SwiftData's ModelContext, FetchDescriptor, and
#Predicate. Each converted service retains a deprecated ModelContext init that
bridges to the shared CoreDataStack.viewContext (both contexts share the same
SQLite store) so unconverted callers continue to compile.

Services converted:
- WorkRepository, WorkCheckInService, WorkStepService, WorkCompletionService
- WorkCompletionBackfill, WorkPresentationStatusService, SampleWorkService
- SchoolDayChecker, PracticeSessionRepository, PracticeSessionMigration
- GroupPracticeHelper, GroupTrackService, RelationshipBackfillService
- WorkAging (primary methods), AttendanceLookup, PresentationAbsentHelper

Bridge patterns used for unconverted view/VM callers:
- Deprecated ModelContext inits → AppBootstrapping.getSharedCoreDataStack().viewContext
- Service protocol adapters (WorkCheckInServiceAdapter, WorkStepServiceAdapter)
  accept SwiftData types, look up CD equivalents by ID, delegate to CD service
- SampleWorkService deprecated overloads accept SampleWork/SampleWorkStep/Lesson
- Views creating entities (PracticeSessionSheet, etc.) bypass CD repo and create
  via SwiftData directly; CD side sees them on next fetch
- Functions returning WorkModel? from CD creation refactored to handle completion
  inline or re-fetch via SwiftData after CD save

+++ Maria's Notebook/Work/SampleWorkEditorSheet.swift
@@ -285,7 +285,7 @@ struct SampleWorkEditorSheet: View {
                 } else {
                     let newStep = service.createStep(for: existing, title: draft.title.trimmed(),
                                                      instructions: draft.instructions)
-                    newStep.orderIndex = index
+                    newStep.orderIndex = Int64(index)
                 }
             }
         } else {
@@ -295,7 +295,7 @@ struct SampleWorkEditorSheet: View {
             for (index, draft) in draftSteps.enumerated() {
                 let step = service.createStep(for: sw, title: draft.title.trimmed(),
                                                instructions: draft.instructions)
-                step.orderIndex = index
+                step.orderIndex = Int64(index)
             }
         }
